Stata_sample_work_project
Over the course of three years I became an expert in using Stata to perform economic analysis on geographic markets. This piece of code was the backbone of the analysis which led to the classification of areas with different competitive conditions. The outputs shaped regulation of telecommunications markets in the UK with £bn consequences for investors and networks.

I led on many areas of the coding and analysis. I managed the work of three junior analysts and I reported to one senior analyst.