Python_blackjack
This game was my second milestone project in the Udemy Python bootcamp. I did not consult the walkthrough Noetbook. It makes use of various elements of Object Oriented Programming including inheritance. It functions as a simplified version of the game with the human player only being able to stick or hit. Aces are only high but I have left some wiggle room for inclusion of aces being high or low.